data/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
playwright>=1.40
pydantic>=2.5
anthropic>=0.40
sqlite-utils>=3.36
python-Levenshtein>=0.25
geopy>=2.4
schedule>=1.2
python-dotenv>=1.0
